# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Enrichment API",
    "description": "API for file enrichment services",
    "version": "0.1.0"
  },
  "paths": {
...
    "/enrichments": {
      "get": {
        "tags": [
          "enrichments"
        ],
        "summary": "List enrichment modules",
        "description": "Get a list of all available enrichment modules",
        "operationId": "list_enrichments_enrichments_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnrichmentsListResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/enrichments/{enrichment_name}": {
      "post": {
        "tags": [
          "enrichments"
        ],
        "summary": "Run enrichment module",
        "description": "Run a specific enrichment module on a file",
        "operationId": "run_enrichment_enrichments__enrichment_name__post",
        "parameters": [
          {
            "name": "enrichment_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Name of the enrichment module to run",
              "title": "Enrichment Name"
            },
            "description": "Name of the enrichment module to run"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnrichmentRequest",
                "description": "The enrichment request containing the object ID"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnrichmentResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EnrichmentRequest": {
        "properties": {
          "object_id": {
            "type": "string",
            "title": "Object Id"
          }
        },
        "type": "object",
        "required": [
          "object_id"
        ],
        "title": "EnrichmentRequest"
      },
      "EnrichmentResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "instance_id": {
            "type": "string",
            "title": "Instance Id"
          },
          "object_id": {
            "type": "string",
            "title": "Object Id"
          }
        },
        "type": "object",
        "required": [
          "status",
          "message",
          "instance_id",
          "object_id"
        ],
        "title": "EnrichmentResponse"
      },
      "EnrichmentsListResponse": {
        "properties": {
          "modules": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Modules"
          }
        },
        "type": "object",
        "required": [
          "modules"
        ],
        "title": "EnrichmentsListResponse"
      },
      "ErrorResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail",
            "description": "Error message details"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  },
...
}